\name{datetime}
\alias{datetime}
\title{datetime}
\description{Construct an Ovation timestamp from date components.}
\usage{
datetime(year,
month,
day,
hour = 0,
minute = 0,
second = 0,
millisecond = 0,
timezone = character())
}
\arguments{
  \item{year}{
}
  \item{month}{
}
  \item{day}{
}
  \item{hour}{
}
  \item{minute}{
}
  \item{second}{
}
  \item{millisecond}{
}
  \item{timezone}{
}
}

\value{A DateTime object}

\author{Romain Franconville}





