\name{EditQuery}
\alias{EditQuery}
\title{Edits an Ovation query using the GUI query editor}
\usage{
  EditQuery(expression.tree = .jnull(class = "com/physion/ebuilder/expression/ExpressionTree"))
}
\arguments{
  \item{expression.tree}{existing expression tree to edit
  (optional)}
}
\value{
  edited expression tree
}
\description{
  Edits an Ovation query using the GUI query editor
}

