\name{List2Map}
\alias{List2Map}
\title{Creates a \code{java.util.Map} from the contents of an R list}
\usage{
  List2Map(l)
}
\arguments{
  \item{l}{R list}
}
\value{
  \code{Map} with the same keys and values as \code{l}
}
\description{
  Creates a \code{java.util.Map} from the contents of an R
  list
}

