\name{newDataContext}
\alias{newDataContext}
\title{newDataContext}
\description{Returns a DataContext object after prompting for password}
\usage{
newDataContext(connection,
login)
}
\arguments{
  \item{connection}{Connection to the database}
  \item{login}{User login}
}

\value{A reference to a DataContext object}

\author{Romain Franconville}





