\name{NewUrl}
\alias{NewUrl}
\title{Creates a new \code{java.net.URL} from the given local path}
\usage{
  NewUrl(path)
}
\arguments{
  \item{path}{local file path (absolute or relative to the
  current working directory)}
}
\value{
  newly created \code{java.net.URL}
}
\description{
  Creates a new \code{java.net.URL} from the given local
  path
}

