\name{nextJavaIt}
\alias{nextJavaIt}
\title{nextJavaIt}
\description{Wrapper for applying the "next" method to an iterator, because "next" is a protected word in R}
\usage{
nextJavaIt(iterator)
}
\arguments{
  \item{iterator}{
}
}

\value{Returns the next object the iterator is refering to }

\author{Romain Franconville}





