\name{NumericData}
\alias{NumericData}
\title{Constructs an new Ovation \code{NumericData} object from an R vector}
\usage{
  NumericData(name, vec, units, sampling.rate,
    sampling.rate.units)
}
\arguments{
  \item{name}{data name}

  \item{vec}{vector of numeric data contents}

  \item{units}{data units}

  \item{sampling.rate}{data sampling rate}

  \item{sampling.rate.units}{data sampling rate units (e.g.
  "Hz")}
}
\value{
  newly constructed \code{NumericData}
}
\description{
  Constructs an new Ovation \code{NumericData} object from
  an R vector
}

