\name{Rovation-package}
\alias{Rovation-package}
\docType{package}
\title{R Interface to the Ovation Scientific Data Management System}
\description{A set of utility allowing the user to access and modify an Ovation database (http://physionconsulting.com/web/Ovation.html)}
\details{
\tabular{ll}{Package: \tab Rovation\cr
Type: \tab Package\cr
Title: \tab R Interface to the Ovation Scientific Data Management System\cr
Version: \tab 0.0.1\cr
Date: \tab 2012-05-08\cr
Author: \tab Romain Franconville\cr
Depends: \tab rJava,tcltk\cr
SystemRequirements : \tab Java\cr
Maintainer: \tab Romain Franconville <franconviller@janelia.hhmi.org>\cr
License: \tab BSD\cr
LazyLoad: \tab yes\cr}
}
\author{Romain Franconville}

\keyword{ package }


