\name{Datetime}
\alias{Datetime}
\title{Datetime}
\description{Construct an Ovation timestamp from date components.}
\usage{Datetime(year, month, day, hour = 0, minute = 0, second = 0, 
    millisecond = 0, timezone = character())}
\arguments{
  \item{year}{
}
  \item{month}{
}
  \item{day}{
}
  \item{hour}{
}
  \item{minute}{
}
  \item{second}{
}
  \item{millisecond}{
}
  \item{timezone}{
}
}



\author{Romain Franconville}





