\name{Rovation-package}
\alias{Rovation-package}
\docType{package}
\title{R Interface to the Ovation Scientific Data Management System}
\description{R API for the Ovation Scientific Data Management System (http://ovation.io)}
\details{
\tabular{ll}{Package: \tab Rovation\cr
Type: \tab Package\cr
Title: \tab R Interface to the Ovation Scientific Data Management System\cr
Version: \tab 2.0-beta2\cr
Date: \tab 2013-07-01\cr
Author: \tab Romain Franconville, Barry Wark\cr
Depends: \tab rJava,tcltk\cr
SystemRequirements : \tab Java\cr
Maintainer: \tab Physion Consulting <support@ovation.io>\cr
License: \tab BSD\cr
LazyLoad: \tab yes\cr}
}
\author{Physion Consulting}

\keyword{ package }


